Rename BudgetSummary to BudgetSummaryPage
